Revert "From Jake: merge test #6"
